Update events.md
If you are not already familiar with the notion of Modules and Plugins, it is not currently clear that using Events requires you develop a plugin or module. Pointing to related documentation to learn more about these related concepts is more helpful than just assuming developers already have that knowledge, or leaving developers confused about the context of this documentation.